requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
mlflow
matplotlib
seaborn
plotly
streamlit
imbalanced-learn
scipy
statsmodels
yellowbrick
joblib
optuna
hyperopt
tune-sklearn
pydantic
scikit-optimize
shap